requirements.txt
# Core dependencies
python-dotenv>=0.19.0
requests>=2.25.0
pandas>=1.3.0
numpy>=1.20.0

# Web crawling and analysis
beautifulsoup4>=4.9.0
selenium>=4.0.0
webdriver-manager>=3.5.0

# API integrations
google-api-python-client>=2.0.0

# NLP and semantic analysis
nltk>=3.6.0

# Data visualization and reporting
matplotlib>=3.4.0

# Web application
flask>=2.0.0
flask-cors>=3.0.0
jinja2>=3.0.0

# Testing
pytest>=6.0.0
